requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via uvicorn
bcrypt==3.2.0
    # via passlib
cffi==1.14.6
    # via
    #   bcrypt
    #   cryptography
click==8.0.1
    # via uvicorn
cryptography==35.0.0
    # via -r requirements.in
fastapi==0.68.2
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via fastapi
pyjwt==2.2.0
    # via -r requirements.in
python-decouple==3.5
    # via -r requirements.in
python-dotenv==0.19.2
    # via uvicorn
pyyaml==6.0
    # via uvicorn
six==1.16.0
    # via bcrypt
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.2
    # via pydantic
uvicorn[standard]==0.15.0
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.0
    # via uvicorn
